Shrink table of contents header
We aren't gonna need 2^32 uids/gids, this is a package manager, not a
file system. Also, mode only uses 16 of 32 bits. We can shrink the
entire header to 8 bytes.

+++ include/pkg/pkgformat.h
@@ -36,10 +36,9 @@ typedef struct {
 } record_t;
 
 typedef struct {
-	uint32_t mode;
-	uint32_t uid;
-	uint32_t gid;
-
+	uint16_t mode;
+	uint16_t uid;
+	uint16_t gid;
 	uint16_t path_length;
 	/* uint8_t path[]; */
 } toc_entry_t;
